added uml and floating point optimized version and some assembly files and I forget what's new and what isn't

+++ Submission/c_src/cordic.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 
-int16_t arctan_fraction_float(int16_t, int16_t);
-int16_t arctan_float(int16_t);
-int16_t cos_float(int32_t angle);
-int16_t sin_float(int32_t angle);
+float arctan_fraction_float(int16_t, int16_t);
+float arctan_float(int16_t);
+float cos_float(int32_t angle);
+float sin_float(int32_t angle);
 						
 int16_t arctan_fraction_routine(int16_t, int16_t);
 int16_t arctan_routine(int16_t);
